RAM usage optimizations
Reduced the size of several fields of the Record structure from uin32_t
to uint16_t or uint8_t. The APIs still take uint32_t, so the internal
representation can be made configurable in the future if larger counters
are required on some platforms.
Updated the API to accept const arrays of pointers to const strings,
so the application can keep both the strings and the pointers to the
strings in read-only memory.

+++ tests/TestFaultInjection.cpp
@@ -93,7 +93,7 @@ static Manager &GetTestFIMgr(void)
         memset(&sFaultAArguments, 0, sizeof(sFaultAArguments));
         sFaultRecordArray[kTestFaultInjectionID_A].mArguments = sFaultAArguments;
         sFaultRecordArray[kTestFaultInjectionID_A].mLengthOfArguments =
-            static_cast<uint16_t>(sizeof(sFaultAArguments)/sizeof(sFaultAArguments[0]));
+            static_cast<uint8_t>(sizeof(sFaultAArguments)/sizeof(sFaultAArguments[0]));
     }
     return sTestFaultInMgr;
 }
